+++ src/common/components/Navigation/SideBar/SideBar.tsx
@@ -1,6 +1,6 @@
 import { useAuth } from "@/application/hooks";
 import { instagramUrl, twitterUrl } from "@/common/utils";
-import { useTagCategoryContext } from "@/components/tags";
+import { useTagCategoryContext } from "@/features/common";
 
 import { Drawer } from "../../Drawer";
 import { Icon } from "../../Icon";
