--- src/components/explore/EmptyMemesView/index.ts
@@ -1,1 +1,0 @@
-export * from "./EmptyMemesView";
