--- src/components/explore/MemesByTag/index.ts
@@ -1,1 +1,0 @@
-export * from "./MemesByTag";
