--- src/components/explore/Thumbnail/index.ts
@@ -1,1 +1,0 @@
-export * from "./Thumbnail";
