--- src/components/explore/index.ts
@@ -1,4 +1,0 @@
-export * from "./EmptyMemesView";
-export * from "./MemesByKeyword";
-export * from "./MemesByTag";
-export * from "./Thumbnail";
