--- src/components/tags/TagFavoriteButton/index.ts
@@ -1,1 +1,0 @@
-export * from "./TagBookmarkButton";
