--- src/components/tags/index.ts
@@ -1,2 +1,0 @@
-export * from "./TagCategory";
-export * from "./TagFavoriteButton";
