+++ src/features/explore/tags/components/index.ts
@@ -1,0 +1,3 @@
+export * from "./MemesByTag";
+export * from "./TagBookmarkButton";
+export * from "./Thumbnail";
