+++ src/pages/_app.tsx
@@ -12,7 +12,7 @@ import { SignUpModal, SignUpModalProvider } from "@/common/components/Modal";
 import { ToastContainer, ToastProvider } from "@/common/components/Toast";
 import { OverlayProvider, RouteTrackingProvider, useAnalytics } from "@/common/hooks";
 import { GoogleTagManagerScript, GTagScript } from "@/common/libs";
-import { TagCategoryProvider } from "@/components/tags";
+import { TagCategoryProvider } from "@/features/common";
 import type { DefaultPageProps } from "@/types";
 
 if (process.env.NEXT_PUBLIC_API_MOCKING === "enabled") {
